Void Functions!
This program introduces the idea of void functions. The program is similar to the program before in which the program returns 10^2, and allows the user to input a base and exponential to calculate the power. This time the use of void functions are apparent which calculates everything without needing to be returned into memory.