Basic Data Types
This program lists all the basic datatypes and explains the uses of them. Furthermore, the program explains in detail Integers and characters. The code explains all the different types of integers and characters, stating the amount of bytes each type of datatype occupies. Also, the Climits library is used to show the minimum and maximum value a datatype can occupy. The program also elaborates on ASCII, and how characters can either store literal numbers, or an Letter using the ASCII equivalent.